refactor Artist objects to have a user object, and have a profile information object. createVenueProfile method now breaking.
