+++ app/src/main/java/com/example/testandroidapplication/utils/HttpJsonParser.java
@@ -137,14 +137,13 @@ public class HttpJsonParser {
         jsonObject.accumulate("User_Id", venue.getUser().getUserID());
         jsonObject.accumulate("User_Name", venue.getUser().getName());
         jsonObject.accumulate("Email", venue.getUser().getEmail());
-        jsonObject.accumulate("Profile_Picture", venue.getUser().getProfileImage());
-        jsonObject.accumulate("Facebook", venue.getUser().getFacebookLink());
-        jsonObject.accumulate("Instagram", venue.getUser().getInstagramLink());
-        jsonObject.accumulate("Twitter", venue.getUser().getTwitterLink());
-        jsonObject.accumulate("Website", venue.getUser().getWebPageLink());
-        jsonObject.accumulate("Tagline", venue.getUser().getTagLine());
-        jsonObject.accumulate("Description", venue.getUser().getTagLine());
-        jsonObject.accumulate("Location", venue.getUser().getLocation());
+        jsonObject.accumulate("Facebook", venue.getProfileInformation().getFacebookLink());
+        jsonObject.accumulate("Instagram", venue.getProfileInformation().getInstagramLink());
+        jsonObject.accumulate("Twitter", venue.getProfileInformation().getTwitterLink());
+        jsonObject.accumulate("Website", venue.getProfileInformation().getWebPageLink());
+        jsonObject.accumulate("Tagline", venue.getProfileInformation().getTagLine());
+        jsonObject.accumulate("Description", venue.getProfileInformation().getTagLine());
+        jsonObject.accumulate("Location", venue.getProfileInformation().getLocation());
         jsonObject.accumulate("Address1", venue.getAddress1());
         jsonObject.accumulate("PostCode", venue.getPostcode());
         jsonObject.accumulate("Phone_Number", venue.getPhoneNumber());
